+++ .github/workflows/performance-base-run-bulk-upload.yml
@@ -1,4 +1,4 @@
-name: "Updated Base Bulk Upload - Execute a Bulk Upload"
+name: "Performance Base Bulk Upload - Execute a Bulk Upload"
 
 permissions:
   pull-requests: write
@@ -71,7 +71,7 @@ jobs:
 
       - name: Setup Bulk Upload
         run: |
-          python updated_setup_bulk_upload.py \
+          python performance_setup_bulk_upload.py \
             --environment "${{ inputs.sandbox }}" \
             --delete-table \
             --download-data \
