+++ .github/workflows/performance-run-bulk-upload-test.yml
@@ -1,4 +1,4 @@
-name: Updated Run Bulk Upload - Test
+name: Performance Run Bulk Upload - Test
 
 permissions:
   pull-requests: write
@@ -31,7 +31,7 @@ on:
 
 jobs:
   bulk_upload:
-    uses: ./.github/workflows/updated-base-run-bulk-upload.yml
+    uses: performance-base-run-bulk-upload.yml
     with:
       environment: development
       sandbox: "${{ inputs.sandbox }}"
