+++ tests/bulk-upload/scripts/performance_setup_bulk_upload.py
